fix: Car is stopping at red lights and starting again on green lights
The issue was due to the decelerate_wp list. Apparently the linear velocity is being set to 0.0 after publish so it was making the car to stop as soon as the lane was calculated (very early before the desired stopping waypoint).
Now we are calculating this deceleration wp list at every loop.